+++ src/main/java/org/truffleruby/language/objects/classvariables/ReadClassVariableNode.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, 2020 Oracle and/or its affiliates. All rights reserved. This
+ * Copyright (c) 2013, 2021 Oracle and/or its affiliates. All rights reserved. This
  * code is released under a tri EPL/GPL/LGPL license. You can use it,
  * redistribute it and/or modify it under the terms of the:
  *
@@ -7,11 +7,10 @@
  * GNU General Public License version 2, or
  * GNU Lesser General Public License version 2.1.
  */
-package org.truffleruby.language.objects;
+package org.truffleruby.language.objects.classvariables;
 
 import org.truffleruby.RubyContext;
 import org.truffleruby.RubyLanguage;
-import org.truffleruby.core.module.ModuleOperations;
 import org.truffleruby.core.module.RubyModule;
 import org.truffleruby.language.LexicalScope;
 import org.truffleruby.language.RubyContextSourceNode;
@@ -29,6 +28,9 @@ public class ReadClassVariableNode extends RubyContextSourceNode {
     private final BranchProfile missingProfile = BranchProfile.create();
 
     @Child private RubyNode lexicalScopeNode;
+    @Child private ResolveTargetModuleForClassVariablesNode resolveTargetModuleNode = ResolveTargetModuleForClassVariablesNode
+            .create();
+    @Child private LookupClassVariableNode lookupClassVariableNode = LookupClassVariableNode.create();
     @Child private WarnNode warnNode;
 
     public ReadClassVariableNode(RubyNode lexicalScopeNode, String name) {
@@ -39,10 +41,8 @@ public class ReadClassVariableNode extends RubyContextSourceNode {
     @Override
     public Object execute(VirtualFrame frame) {
         final LexicalScope lexicalScope = (LexicalScope) lexicalScopeNode.execute(frame);
-        // TODO CS 21-Feb-16 these two operations are uncached and use loops - same for isDefined below
-        final RubyModule module = LexicalScope.resolveTargetModuleForClassVariables(lexicalScope);
-
-        final Object value = ModuleOperations.lookupClassVariable(module, name);
+        final RubyModule module = resolveTargetModuleNode.execute(lexicalScope);
+        final Object value = lookupClassVariableNode.execute(module, name);
 
         if (value == null) {
             missingProfile.enter();
@@ -61,9 +61,8 @@ public class ReadClassVariableNode extends RubyContextSourceNode {
     @Override
     public Object isDefined(VirtualFrame frame, RubyLanguage language, RubyContext context) {
         final LexicalScope lexicalScope = (LexicalScope) lexicalScopeNode.execute(frame);
-        final RubyModule module = LexicalScope.resolveTargetModuleForClassVariables(lexicalScope);
-
-        final Object value = ModuleOperations.lookupClassVariable(module, name);
+        final RubyModule module = resolveTargetModuleNode.execute(lexicalScope);
+        final Object value = lookupClassVariableNode.execute(module, name);
 
         if (lexicalScope.getParent() == null) {
             warnTopLevelClassVariableAccess();